vtt-sync: phase 4 extract board state poller to services/board-state-poller.js
Move the createBoardStatePoller factory (~340 lines) out of
board-interactions.js into its own module. The factory's two interval
constants (POLL_FALLBACK_INTERVAL_MS, POLL_SAFETY_NET_INTERVAL_MS) move
with it. board-interactions.js imports the factory and re-exports it so
existing tests that import createBoardStatePoller from
board-interactions.js continue to work.

Behavior unchanged. The poller's 24 existing tests still pass without
modification. board-interactions.js went from 20,806 to 20,468 lines.
